Added additional checks for fpin events.
Incorporated the review comments by Benjamin.

Enabled debug logs in Makefile

Made changes in the code to set path to marginal instead of
fail path.

Added additional checks for the FPIN events and ignore the other
events which are not supported at present with the log.

Added changes to unset the marginal paths on receving
LINKUP and RSCN events

Updated the README doc accordingly